python-services/rag/requirements.txt
# RAG Service Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector Database
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.2.2

# Database
supabase>=2.0.0

# Google Gemini for reranking/generation
google-genai

# Text Processing
langchain-text-splitters>=0.0.1

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.0
